Makefile: Fix a warning related to use write syscall
When I built the program on ubuntu 16.04, I got a warning related to write syscall.
It looks like:
warning: ignoring return value of ‘write’, declared with attribute warn_unused_result [-Wunused-result]

We did not check the return value of using write syscall. So, add a more flag for ignoring
the warning related to unused result.

Signed-off-by: Hoang Van Tuyen <tuyen.hv.3i@gmail.com>